test: 테스트 코드 파일 main 폴더 구조와 유사하게 변경
