--- src/test/java/lotto/WinningRank.java
@@ -1,4 +1,0 @@
-package lotto;
-
-public class WinningRank {
-}
