+++ src/test/java/lotto/domain/LottoNumberTest.java
@@ -1,11 +1,9 @@
-package lotto;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.List;
-import lotto.domain.Lotto;
-import lotto.domain.LottoMachine;
 import org.junit.jupiter.api.Test;
 
 public class LottoNumberTest {
@@ -19,6 +17,7 @@ public class LottoNumberTest {
         long duplicateCount = ticket.stream().distinct().count();
         assertEquals(6, duplicateCount);
     }
+
     //2.6개로 제대로 나왔는가
     @Test
     void 지정한_로또_크기만큼_나왔는가() {
@@ -27,6 +26,7 @@ public class LottoNumberTest {
         List<Integer> ticket = tickets.get(0);
         assertEquals(6, ticket.size());
     }
+
     //3.정렬이 되어있는가
     @Test
     void 정렬이_되었는가() {
