+++ src/test/java/lotto/integration/LottoTotalTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.integration;
 
 import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomUniqueNumbersInRangeTest;
 import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
@@ -6,6 +6,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import camp.nextstep.edu.missionutils.test.NsTest;
 import java.util.List;
+import lotto.Application;
 import org.junit.jupiter.api.Test;
 
 public class LottoTotalTest extends NsTest {
