+++ src/test/java/lotto/service/WinningNumberUnitTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -8,25 +8,24 @@ import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
 import java.util.List;
 import lotto.exception.LottoMessage;
-import lotto.service.LottoInputService;
 import lotto.util.CountConverter;
 import lotto.util.NumbersConverter;
 import lotto.view.InputView;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 
-public class BonusUnitTest {
+public class WinningNumberUnitTest {
+
     @AfterEach
     void afterEach() {
         Console.close(); // 리소스 해제
     }
 
     @Test
-    void 보너스번호_정상_입력() {
+    void 당첨번호_사용자_정상_입력_시() {
         //given
-        List<Integer> winningNumbers = List.of(1, 2, 3, 4, 5, 6);
-        String bonusInput = "37\n";
-        System.setIn(new ByteArrayInputStream(bonusInput.getBytes()));
+        String winningInput = "1,2,3,4,5,6\n";
+        System.setIn(new ByteArrayInputStream(winningInput.getBytes()));
 
         InputView inputView = new InputView();
 
@@ -37,18 +36,17 @@ public class BonusUnitTest {
         );
 
         //when
-        int bonus = service.checkBonus(winningNumbers);
+        List<Integer> winningNumbers = service.checkNumbers();
 
         //then
-        assertThat(bonus).isEqualTo(37);
+        assertThat(winningNumbers).containsExactly(1, 2, 3, 4, 5, 6);
     }
 
     @Test
-    void 보너스번호_문자_입력시_에러발생() {
+    void 당첨번호_문자_입력시_에러발생() {
         //given
-        List<Integer> winningNumbers = List.of(1, 2, 3, 4, 5, 6);
-        String bonusInput = "!\na\n7";
-        System.setIn(new ByteArrayInputStream(bonusInput.getBytes()));
+        String winningInput = "1,2,3,@,5,6\n1,2,3,4,5,6";
+        System.setIn(new ByteArrayInputStream(winningInput.getBytes()));
 
         InputView inputView = new InputView();
 
@@ -62,19 +60,18 @@ public class BonusUnitTest {
         System.setOut(new PrintStream(out));
 
         //when
-        service.checkBonus(winningNumbers);
+        service.checkNumbers();
 
-        //then
+        // then
         assertThat(out.toString()).contains("[ERROR]");
-        assertThat(out.toString()).contains(LottoMessage.BONUS_MESSAGE);
+        assertThat(out.toString()).contains(LottoMessage.WINNING_MESSAGE);
     }
 
     @Test
-    void 보너스번호_빈_문자열시_에러발생() {
+    void 당첨번호_빈문자열_시_에러발생() {
         //given
-        List<Integer> winningNumbers = List.of(1, 2, 3, 4, 5, 6);
-        String bonusInput = "\n \n7";
-        System.setIn(new ByteArrayInputStream(bonusInput.getBytes()));
+        String winningInput = " \n1,2,3,4,5,6\n";
+        System.setIn(new ByteArrayInputStream(winningInput.getBytes()));
 
         InputView inputView = new InputView();
 
@@ -88,19 +85,18 @@ public class BonusUnitTest {
         System.setOut(new PrintStream(out));
 
         //when
-        service.checkBonus(winningNumbers);
+        service.checkNumbers();
 
-        //then
+        // then
         assertThat(out.toString()).contains("[ERROR]");
-        assertThat(out.toString()).contains(LottoMessage.BONUS_MESSAGE);
+        assertThat(out.toString()).contains(LottoMessage.WINNING_MESSAGE);
     }
 
     @Test
-    void 보너스번호_범위_벗어날경우_에러발생() {
+    void 당첨번호_5개만_입력시_에러발생() {
         //given
-        List<Integer> winningNumbers = List.of(1, 2, 3, 4, 5, 6);
-        String bonusInput = "57\n7";
-        System.setIn(new ByteArrayInputStream(bonusInput.getBytes()));
+        String winningInput = "1,2,3,4,5\n1,2,3,4,5,6";
+        System.setIn(new ByteArrayInputStream(winningInput.getBytes()));
 
         InputView inputView = new InputView();
 
@@ -114,19 +110,18 @@ public class BonusUnitTest {
         System.setOut(new PrintStream(out));
 
         //when
-        service.checkBonus(winningNumbers);
+        service.checkNumbers();
 
-        //then
+        // then
         assertThat(out.toString()).contains("[ERROR]");
-        assertThat(out.toString()).contains(LottoMessage.BONUS_MESSAGE);
+        assertThat(out.toString()).contains(LottoMessage.WINNING_MESSAGE);
     }
 
     @Test
-    void 보너스번호_당첨번호와_중복_에러발생() {
+    void 당첨번호_중복_입력시_에러발생() {
         //given
-        List<Integer> winningNumbers = List.of(1, 2, 3, 4, 5, 6);
-        String bonusInput = "2\n7";
-        System.setIn(new ByteArrayInputStream(bonusInput.getBytes()));
+        String winningInput = "1,2,3,4,5,5\n1,2,3,4,5,6";
+        System.setIn(new ByteArrayInputStream(winningInput.getBytes()));
 
         InputView inputView = new InputView();
 
@@ -140,10 +135,35 @@ public class BonusUnitTest {
         System.setOut(new PrintStream(out));
 
         //when
-        service.checkBonus(winningNumbers);
+        service.checkNumbers();
 
-        //then
+        // then
         assertThat(out.toString()).contains("[ERROR]");
-        assertThat(out.toString()).contains(LottoMessage.BONUS_MESSAGE);
+        assertThat(out.toString()).contains(LottoMessage.WINNING_MESSAGE);
+    }
+
+    @Test
+    void 당첨번호_범위가_아닌경우_에러발생() {
+        //given
+        String winningInput = "1,2,3,4,5,57\n1,2,3,4,5,6";
+        System.setIn(new ByteArrayInputStream(winningInput.getBytes()));
+
+        InputView inputView = new InputView();
+
+        LottoInputService service = new LottoInputService(
+                inputView,
+                new CountConverter(),
+                new NumbersConverter()
+        );
+
+        ByteArrayOutputStream out = new ByteArrayOutputStream();
+        System.setOut(new PrintStream(out));
+
+        //when
+        service.checkNumbers();
+
+        // then
+        assertThat(out.toString()).contains("[ERROR]");
+        assertThat(out.toString()).contains(LottoMessage.WINNING_MESSAGE);
     }
 }
