+++ src/test/java/lotto/util/WinningNumberTest.java
@@ -1,10 +1,9 @@
-package lotto;
+package lotto.util;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import lotto.exception.NumberErrorMessage;
-import lotto.util.NumbersConverter;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
